api/requirements.txt
fastapi>=0.115.12
uvicorn>=0.34.2
openai>=1.93.0
python-dotenv>=1.0.0
PyMuPDF>=1.23.0
numpy>=1.24.0
python-multipart>=0.0.6